Extract SlideUtil methods and remove duplicate ShapeUtil methods
