+++ PowerPointLabs/PowerPointLabs/ActionFramework/Common/Interface/EnabledHandler.cs
@@ -1,12 +1,5 @@
 ﻿using System;
 
-using Microsoft.Office.Interop.PowerPoint;
-
-using PowerPointLabs.ActionFramework.Common.Extension;
-using PowerPointLabs.Utils;
-
-using Office = Microsoft.Office.Core;
-
 namespace PowerPointLabs.ActionFramework.Common.Interface
 {
     /// <summary>
@@ -29,140 +22,5 @@ namespace PowerPointLabs.ActionFramework.Common.Interface
         }
 
         protected abstract bool GetEnabled(string ribbonId);
-        
-        protected bool HasPlaceholderInSelection()
-        {
-            Selection selection = this.GetCurrentSelection();
-            
-            if (selection.Type != PpSelectionType.ppSelectionShapes)
-            {
-                return false;
-            }
-
-            foreach (Shape shape in selection.ShapeRange)
-            {
-                if (shape.Type == Microsoft.Office.Core.MsoShapeType.msoPlaceholder)
-                {
-                    return true;
-                }
-            }
-
-            return false;
-        }
-
-        protected bool IsSelectionSingleShape()
-        {
-            Selection selection = this.GetCurrentSelection();
-
-            if (selection.HasChildShapeRange)
-            {
-                return selection.ChildShapeRange.Count == 1 &&
-                    selection.ChildShapeRange.Type == Microsoft.Office.Core.MsoShapeType.msoAutoShape;
-            }
-
-            return selection.Type == PpSelectionType.ppSelectionShapes && 
-                selection.ShapeRange.Count == 1;
-        }
-
-        protected bool IsSelectionMultipleOrGroup()
-        {
-            Selection selection = this.GetCurrentSelection();
-
-            if (selection.Type != PpSelectionType.ppSelectionShapes)
-            {
-                return false;
-            }
-
-            if (selection.ShapeRange.Count > 1)
-            {
-                return true;
-            }
-
-            if (ShapesUtil.IsAGroup(selection.ShapeRange[1]))
-            {
-                return true;
-            }
-
-            return false;
-        }
-
-        protected bool IsSelectionMultipleSameShapeType()
-        {
-            bool isEnabled = false;
-            var selection = this.GetCurrentSelection();
-            if (selection.Type != PpSelectionType.ppSelectionShapes)
-            {
-                return isEnabled;
-            }
-
-            var shape = selection.ShapeRange[1];
-
-            if (selection.ShapeRange.Count > 1)
-            {
-                foreach (Shape tempShape in selection.ShapeRange)
-                {
-                    if (shape.Type == tempShape.Type)
-                    {
-                        isEnabled = true;
-                    }
-                    if (shape.Type == Microsoft.Office.Core.MsoShapeType.msoAutoShape &&
-                        shape.AutoShapeType != tempShape.AutoShapeType)
-                    {
-                        isEnabled = false;
-                        break;
-                    }
-                }
-            }
-            return isEnabled;
-        }
-
-        protected bool IsSelectionAllRectangle()
-        {
-            Selection selection = this.GetCurrentSelection();
-            if (selection.Type != PpSelectionType.ppSelectionShapes)
-            {
-                return false;
-            }
-            
-            ShapeRange shapes = selection.ShapeRange;
-            foreach (Shape shape in shapes)
-            {
-                if ((shape.Type != Microsoft.Office.Core.MsoShapeType.msoAutoShape ||
-                   shape.AutoShapeType != Microsoft.Office.Core.MsoAutoShapeType.msoShapeRectangle) &&
-                   shape.Type != Microsoft.Office.Core.MsoShapeType.msoPicture)
-                {
-                    return false;
-                }
-            }
-            return true;
-        }
-
-        protected bool IsSelectionAllShapeWithArea()
-        {
-            Selection selection = this.GetCurrentSelection();
-            if (selection.Type != PpSelectionType.ppSelectionShapes)
-            {
-                return false;
-            }
-
-            ShapeRange shapes = selection.ShapeRange;
-            foreach (Shape shape in shapes)
-            {
-                if (shape.Type != Office.MsoShapeType.msoAutoShape && shape.Type != Office.MsoShapeType.msoFreeform &&
-                    shape.Type != Office.MsoShapeType.msoTextBox && shape.Type != Office.MsoShapeType.msoPlaceholder &&
-                    shape.Type != Office.MsoShapeType.msoCallout && shape.Type != Office.MsoShapeType.msoInk &&
-                    shape.Type != Office.MsoShapeType.msoGroup)
-                {
-                    return false;
-                }
-            }
-            return true;
-        }
-
-        protected bool IsSelectionChildShapeRange()
-        {
-            Selection selection = this.GetCurrentSelection();
-            return selection.HasChildShapeRange;
-        }
     }
 }
