+++ PowerPointLabs/PowerPointLabs/ActionFramework/EffectsLab/Spotlight/AddSpotlight/AddSpotlightEnabledHandler.cs
@@ -1,5 +1,9 @@
-﻿using PowerPointLabs.ActionFramework.Common.Attribute;
+﻿using Microsoft.Office.Interop.PowerPoint;
+
+using PowerPointLabs.ActionFramework.Common.Attribute;
+using PowerPointLabs.ActionFramework.Common.Extension;
 using PowerPointLabs.ActionFramework.Common.Interface;
+using PowerPointLabs.Utils;
 
 namespace PowerPointLabs.ActionFramework.EffectsLab
 {
@@ -8,7 +12,8 @@ namespace PowerPointLabs.ActionFramework.EffectsLab
     {
         protected override bool GetEnabled(string ribbonId)
         {
-            return IsSelectionAllShapeWithArea();
+            Selection currentSelection = this.GetCurrentSelection();
+            return ShapeUtil.IsSelectionAllShapeWithArea(currentSelection);
         }
     }
 }
