+++ PowerPointLabs/PowerPointLabs/ActionFramework/PasteLab/PasteIntoGroup/PasteIntoGroupActionHandler.cs
@@ -20,7 +20,7 @@ namespace PowerPointLabs.ActionFramework.PasteLab
                 return null;
             }
 
-            if (selectedShapes.Count == 1 && !ShapesUtil.IsAGroup(selectedShapes[1]))
+            if (selectedShapes.Count == 1 && !ShapeUtil.IsAGroup(selectedShapes[1]))
             {
                 Logger.Log("PasteIntoGroup failed. Selection is only a single shape.");
                 return null;
