+++ PowerPointLabs/PowerPointLabs/ActionFramework/PasteLab/PasteIntoGroup/PasteIntoGroupEnabledHandler.cs
@@ -1,4 +1,7 @@
-﻿using PowerPointLabs.ActionFramework.Common.Attribute;
+﻿using Microsoft.Office.Interop.PowerPoint;
+
+using PowerPointLabs.ActionFramework.Common.Attribute;
+using PowerPointLabs.ActionFramework.Common.Extension;
 using PowerPointLabs.ActionFramework.Common.Interface;
 using PowerPointLabs.Utils;
 
@@ -9,9 +12,10 @@ namespace PowerPointLabs.ActionFramework.PasteLab
     {
         protected override bool GetEnabled(string ribbonId)
         {
-            return !GraphicsUtil.IsClipboardEmpty() && 
-                IsSelectionMultipleOrGroup() &&
-                !HasPlaceholderInSelection();
+            Selection currentSelection = this.GetCurrentSelection();
+            return !GraphicsUtil.IsClipboardEmpty() &&
+                ShapeUtil.IsSelectionMultipleOrGroup(currentSelection) &&
+                !ShapeUtil.HasPlaceholderInSelection(currentSelection);
         }
     }
 }
