+++ PowerPointLabs/PowerPointLabs/ActionFramework/ShortcutsLab/AddIntoGroup/AddIntoGroupEnabledHandler.cs
@@ -1,5 +1,8 @@
-﻿using PowerPointLabs.ActionFramework.Common.Attribute;
+﻿using Microsoft.Office.Interop.PowerPoint;
+using PowerPointLabs.ActionFramework.Common.Attribute;
+using PowerPointLabs.ActionFramework.Common.Extension;
 using PowerPointLabs.ActionFramework.Common.Interface;
+using PowerPointLabs.Utils;
 
 namespace PowerPointLabs.ActionFramework.ShortcutsLab
 {
@@ -8,7 +11,10 @@ namespace PowerPointLabs.ActionFramework.ShortcutsLab
     {
         protected override bool GetEnabled(string ribbonId)
         {
-            return !IsSelectionSingleShape() && !HasPlaceholderInSelection() && !IsSelectionChildShapeRange();
+            Selection currentSelection = this.GetCurrentSelection();
+            return !ShapeUtil.IsSelectionSingleShape(currentSelection) && 
+                !ShapeUtil.HasPlaceholderInSelection(currentSelection) && 
+                !ShapeUtil.IsSelectionChildShapeRange(currentSelection);
         }
     }
 }
