+++ PowerPointLabs/PowerPointLabs/ActionFramework/ShortcutsLab/EditName/EditNameEnabledHandler.cs
@@ -1,4 +1,7 @@
-﻿using PowerPointLabs.ActionFramework.Common.Attribute;
+﻿using Microsoft.Office.Interop.PowerPoint;
+
+using PowerPointLabs.ActionFramework.Common.Attribute;
+using PowerPointLabs.ActionFramework.Common.Extension;
 using PowerPointLabs.ActionFramework.Common.Interface;
 using PowerPointLabs.Utils;
 
@@ -9,7 +12,8 @@ namespace PowerPointLabs.ActionFramework.ShortcutsLab
     {
         protected override bool GetEnabled(string ribbonId)
         {
-            return IsSelectionSingleShape();
+            Selection currentSelection = this.GetCurrentSelection();
+            return ShapeUtil.IsSelectionSingleShape(currentSelection);
         }
     }
 }
