+++ PowerPointLabs/PowerPointLabs/ActionFramework/ZoomLab/ZoomToArea/ZoomToAreaEnabledHandler.cs
@@ -1,5 +1,9 @@
-﻿using PowerPointLabs.ActionFramework.Common.Attribute;
+﻿using Microsoft.Office.Interop.PowerPoint;
+
+using PowerPointLabs.ActionFramework.Common.Attribute;
+using PowerPointLabs.ActionFramework.Common.Extension;
 using PowerPointLabs.ActionFramework.Common.Interface;
+using PowerPointLabs.Utils;
 
 namespace PowerPointLabs.ActionFramework.AnimationLab
 {
@@ -8,7 +12,8 @@ namespace PowerPointLabs.ActionFramework.AnimationLab
     {
         protected override bool GetEnabled(string ribbonId)
         {
-            return IsSelectionAllRectangle();
+            Selection currentSelection = this.GetCurrentSelection();
+            return ShapeUtil.IsSelectionAllRectangle(currentSelection);
         }
     }
 }
