+++ PowerPointLabs/PowerPointLabs/AnimationLab/AnimateInSlide.cs
@@ -111,7 +111,7 @@ namespace PowerPointLabs.AnimationLab
                         PowerPoint.MsoAnimTriggerType.msoAnimTriggerAfterPrevious);
                 }
                 PowerPoint.Effect shape1Disappear = currentSlide.TimeLine.MainSequence.AddEffect(
-                        Utils.ShapesUtil.IsStraightLine(shape1) ? shape1.ParentGroup : shape1,
+                        Utils.ShapeUtil.IsStraightLine(shape1) ? shape1.ParentGroup : shape1,
                         PowerPoint.MsoAnimEffect.msoAnimEffectFade,
                         PowerPoint.MsoAnimateByLevel.msoAnimateLevelNone,
                         PowerPoint.MsoAnimTriggerType.msoAnimTriggerWithPrevious);
@@ -152,7 +152,7 @@ namespace PowerPointLabs.AnimationLab
 
             if ((AnimationLabSettings.IsUseFrameAnimation && (shape2.Height != shape1.Height || shape2.Width != shape1.Width)) || 
                 ((shape2.Rotation != shape1.Rotation || shape1.Rotation % 90 != 0) && (shape2.Height != shape1.Height || shape2.Width != shape1.Width)) || 
-                (!Utils.ShapesUtil.IsStraightLine(shape1) && (shape1.HorizontalFlip != shape2.HorizontalFlip || shape1.VerticalFlip != shape2.VerticalFlip)) || 
+                (!Utils.ShapeUtil.IsStraightLine(shape1) && (shape1.HorizontalFlip != shape2.HorizontalFlip || shape1.VerticalFlip != shape2.VerticalFlip)) || 
                 finalFont != initialFont)
             {
                 return true;
