+++ PowerPointLabs/PowerPointLabs/Models/PowerPointAutoAnimateSlide.cs
@@ -115,7 +115,7 @@ namespace PowerPointLabs.Models
 
             if ((AnimationLabSettings.IsUseFrameAnimation && (shape2.Height != shape1.Height || shape2.Width != shape1.Width)) || 
                 ((shape2.Rotation != shape1.Rotation || shape1.Rotation % 90 != 0) && (shape2.Height != shape1.Height || shape2.Width != shape1.Width)) || 
-                (!Utils.ShapesUtil.IsStraightLine(shape1) && (shape1.HorizontalFlip != shape2.HorizontalFlip || shape1.VerticalFlip != shape2.VerticalFlip)) || 
+                (!Utils.ShapeUtil.IsStraightLine(shape1) && (shape1.HorizontalFlip != shape2.HorizontalFlip || shape1.VerticalFlip != shape2.VerticalFlip)) || 
                 finalFont != initialFont)
             {
                 return true;
