+++ PowerPointLabs/PowerPointLabs/PositionsLab/PositionsLabMain.cs
@@ -410,14 +410,14 @@ namespace PowerPointLabs.PositionsLab
                 throw new Exception(ErrorMessageFewerThanThreeSelection);
             }
                     
-            var origin = ShapesUtil.GetCenterPoint(selectedShapes[1]);
-            var refPoint = ShapesUtil.GetCenterPoint(selectedShapes[2]);
+            var origin = ShapeUtil.GetCenterPoint(selectedShapes[1]);
+            var refPoint = ShapeUtil.GetCenterPoint(selectedShapes[2]);
             var distance = DistanceBetweenTwoPoints(origin, refPoint);
 
             for (var i = 3; i <= selectedShapes.Count; i++)
             {
                 var shape = selectedShapes[i];
-                var point = ShapesUtil.GetCenterPoint(shape);
+                var point = ShapeUtil.GetCenterPoint(shape);
                 var currentDistance = DistanceBetweenTwoPoints(origin, point);
                 var proportion = (currentDistance - distance) / currentDistance;
 
@@ -437,7 +437,7 @@ namespace PowerPointLabs.PositionsLab
             }
 
             var refShape = selectedShapes[0];
-            var sortedShapes = ShapesUtil.SortShapesByLeft(selectedShapes);
+            var sortedShapes = ShapeUtil.SortShapesByLeft(selectedShapes);
             var refShapeIndex = sortedShapes.IndexOf(refShape);
 
             var mostLeft = refShape.VisualLeft;
@@ -478,7 +478,7 @@ namespace PowerPointLabs.PositionsLab
             }
 
             var refShape = selectedShapes[0];
-            var sortedShapes = ShapesUtil.SortShapesByTop(selectedShapes);
+            var sortedShapes = ShapeUtil.SortShapesByTop(selectedShapes);
             var refShapeIndex = sortedShapes.IndexOf(refShape);
 
             var mostTop = refShape.VisualTop;
@@ -536,7 +536,7 @@ namespace PowerPointLabs.PositionsLab
 
                 startingPoint = 0;
                 referenceWidth = slideWidth;
-                shapesToDistribute = ShapesUtil.SortShapesByLeft(selectedShapes);
+                shapesToDistribute = ShapeUtil.SortShapesByLeft(selectedShapes);
 
                 foreach (var s in shapesToDistribute)
                 {
@@ -588,7 +588,7 @@ namespace PowerPointLabs.PositionsLab
 
                 startingPoint = 0;
                 referenceWidth = slideWidth;
-                shapesToDistribute = ShapesUtil.SortShapesByLeft(selectedShapes);
+                shapesToDistribute = ShapeUtil.SortShapesByLeft(selectedShapes);
 
                 foreach (var s in shapesToDistribute)
                 {
@@ -643,7 +643,7 @@ namespace PowerPointLabs.PositionsLab
                 startingPoint = shapesToDistribute[0].VisualLeft;
                 referenceWidth = shapesToDistribute[0].AbsoluteWidth;
                 shapesToDistribute.RemoveAt(0);
-                shapesToDistribute = ShapesUtil.SortShapesByLeft(shapesToDistribute);
+                shapesToDistribute = ShapeUtil.SortShapesByLeft(shapesToDistribute);
 
                 foreach (var s in shapesToDistribute)
                 {
@@ -697,7 +697,7 @@ namespace PowerPointLabs.PositionsLab
                 startingPoint = shapesToDistribute[0].VisualLeft;
                 referenceWidth = shapesToDistribute[0].AbsoluteWidth;
                 shapesToDistribute.RemoveAt(0);
-                shapesToDistribute = ShapesUtil.SortShapesByLeft(shapesToDistribute);
+                shapesToDistribute = ShapeUtil.SortShapesByLeft(shapesToDistribute);
 
                 foreach (var s in shapesToDistribute)
                 {
@@ -768,7 +768,7 @@ namespace PowerPointLabs.PositionsLab
 
                 shapesToDistribute.RemoveAt(1);
                 shapesToDistribute.RemoveAt(0);
-                shapesToDistribute = ShapesUtil.SortShapesByLeft(shapesToDistribute);
+                shapesToDistribute = ShapeUtil.SortShapesByLeft(shapesToDistribute);
 
                 for (var i = 0; i < shapesToDistribute.Count; i++)
                 {
@@ -815,7 +815,7 @@ namespace PowerPointLabs.PositionsLab
                 shapesToDistribute.RemoveAt(1);
                 shapesToDistribute.RemoveAt(0);
 
-                shapesToDistribute = ShapesUtil.SortShapesByLeft(shapesToDistribute);
+                shapesToDistribute = ShapeUtil.SortShapesByLeft(shapesToDistribute);
 
                 for (var i = 0; i < shapesToDistribute.Count; i++)
                 {
@@ -841,7 +841,7 @@ namespace PowerPointLabs.PositionsLab
                     throw new Exception(ErrorMessageFewerThanThreeSelection);
                 }
 
-                shapesToDistribute = ShapesUtil.SortShapesByLeft(selectedShapes);
+                shapesToDistribute = ShapeUtil.SortShapesByLeft(selectedShapes);
                 startingPoint = shapesToDistribute[0].VisualLeft + shapesToDistribute[0].AbsoluteWidth;
                 var rightMostShape = shapesToDistribute[shapesToDistribute.Count - 1];
                 referenceWidth = rightMostShape.VisualLeft + rightMostShape.AbsoluteWidth - shapesToDistribute[0].VisualLeft;
@@ -879,7 +879,7 @@ namespace PowerPointLabs.PositionsLab
                     throw new Exception(ErrorMessageFewerThanThreeSelection);
                 }
 
-                shapesToDistribute = ShapesUtil.SortShapesByLeft(selectedShapes);
+                shapesToDistribute = ShapeUtil.SortShapesByLeft(selectedShapes);
                 startingPoint = shapesToDistribute[0].VisualCenter.X;
                 referenceWidth = shapesToDistribute[shapesToDistribute.Count-1].VisualCenter.X - shapesToDistribute[0].VisualCenter.X;
 
@@ -932,7 +932,7 @@ namespace PowerPointLabs.PositionsLab
 
                 startingPoint = 0;
                 referenceHeight = slideHeight;
-                shapesToDistribute = ShapesUtil.SortShapesByTop(selectedShapes);
+                shapesToDistribute = ShapeUtil.SortShapesByTop(selectedShapes);
                 foreach (var s in shapesToDistribute)
                 {
                     totalShapeHeight += s.AbsoluteHeight;
@@ -983,7 +983,7 @@ namespace PowerPointLabs.PositionsLab
 
                 startingPoint = 0;
                 referenceHeight = slideHeight;
-                shapesToDistribute = ShapesUtil.SortShapesByTop(selectedShapes);
+                shapesToDistribute = ShapeUtil.SortShapesByTop(selectedShapes);
 
                 foreach (var s in shapesToDistribute)
                 {
@@ -1038,7 +1038,7 @@ namespace PowerPointLabs.PositionsLab
                 startingPoint = shapesToDistribute[0].VisualTop;
                 referenceHeight = shapesToDistribute[0].AbsoluteHeight;
                 shapesToDistribute.RemoveAt(0);
-                shapesToDistribute = ShapesUtil.SortShapesByTop(shapesToDistribute);
+                shapesToDistribute = ShapeUtil.SortShapesByTop(shapesToDistribute);
 
                 foreach (var s in shapesToDistribute)
                 {
@@ -1092,7 +1092,7 @@ namespace PowerPointLabs.PositionsLab
                 startingPoint = shapesToDistribute[0].VisualTop;
                 referenceHeight = shapesToDistribute[0].AbsoluteHeight;
                 shapesToDistribute.RemoveAt(0);
-                shapesToDistribute = ShapesUtil.SortShapesByTop(shapesToDistribute);
+                shapesToDistribute = ShapeUtil.SortShapesByTop(shapesToDistribute);
 
                 foreach (var s in shapesToDistribute)
                 {
@@ -1163,7 +1163,7 @@ namespace PowerPointLabs.PositionsLab
 
                 shapesToDistribute.RemoveAt(1);
                 shapesToDistribute.RemoveAt(0);
-                shapesToDistribute = ShapesUtil.SortShapesByTop(shapesToDistribute);
+                shapesToDistribute = ShapeUtil.SortShapesByTop(shapesToDistribute);
 
                 for (var i = 0; i < shapesToDistribute.Count; i++)
                 {
@@ -1210,7 +1210,7 @@ namespace PowerPointLabs.PositionsLab
                 shapesToDistribute.RemoveAt(1);
                 shapesToDistribute.RemoveAt(0);
 
-                shapesToDistribute = ShapesUtil.SortShapesByTop(shapesToDistribute);
+                shapesToDistribute = ShapeUtil.SortShapesByTop(shapesToDistribute);
 
                 for (var i = 0; i < shapesToDistribute.Count; i++)
                 {
@@ -1236,7 +1236,7 @@ namespace PowerPointLabs.PositionsLab
                     throw new Exception(ErrorMessageFewerThanThreeSelection);
                 }
 
-                shapesToDistribute = ShapesUtil.SortShapesByTop(selectedShapes);
+                shapesToDistribute = ShapeUtil.SortShapesByTop(selectedShapes);
                 startingPoint = shapesToDistribute[0].VisualTop + shapesToDistribute[0].AbsoluteHeight;
                 var bottomMostShape = shapesToDistribute[shapesToDistribute.Count - 1];
                 referenceHeight = bottomMostShape.VisualTop + bottomMostShape.AbsoluteHeight - shapesToDistribute[0].VisualTop;
@@ -1274,7 +1274,7 @@ namespace PowerPointLabs.PositionsLab
                     throw new Exception(ErrorMessageFewerThanThreeSelection);
                 }
 
-                shapesToDistribute = ShapesUtil.SortShapesByTop(selectedShapes);
+                shapesToDistribute = ShapeUtil.SortShapesByTop(selectedShapes);
                 startingPoint = shapesToDistribute[0].VisualCenter.Y;
                 referenceHeight = shapesToDistribute[shapesToDistribute.Count - 1].VisualCenter.Y - shapesToDistribute[0].VisualCenter.Y;
 
@@ -1960,7 +1960,7 @@ namespace PowerPointLabs.PositionsLab
                     throw new Exception(ErrorMessageFewerThanThreeSelection);
                 }
 
-                origin = ShapesUtil.GetCenterPoint(selectedShapes[1]);
+                origin = ShapeUtil.GetCenterPoint(selectedShapes[1]);
 
                 var boundaryAngles = GetShapeBoundaryAngles(origin, selectedShapes[2]);
                 startingAngle = boundaryAngles[1];
@@ -1992,7 +1992,7 @@ namespace PowerPointLabs.PositionsLab
                     throw new Exception(ErrorMessageFewerThanThreeSelection);
                 }
                 
-                origin = ShapesUtil.GetCenterPoint(selectedShapes[1]);
+                origin = ShapeUtil.GetCenterPoint(selectedShapes[1]);
                 
                 startingAngle = (float)AngleBetweenTwoPoints(origin, GetVisualCenter(selectedShapes[2]));
                 referenceAngle = 360;
@@ -2006,7 +2006,7 @@ namespace PowerPointLabs.PositionsLab
                     throw new Exception(ErrorMessageFewerThanFourSelection);
                 }
 
-                origin = ShapesUtil.GetCenterPoint(selectedShapes[1]);
+                origin = ShapeUtil.GetCenterPoint(selectedShapes[1]);
                 var startingShapeBoundaryAngles = GetShapeBoundaryAngles(origin, selectedShapes[2]);
                 var endingShapeBoundaryAngles = GetShapeBoundaryAngles(origin, selectedShapes[3]);
                 startingAngle = startingShapeBoundaryAngles[0];
@@ -2046,7 +2046,7 @@ namespace PowerPointLabs.PositionsLab
                     throw new Exception(ErrorMessageFewerThanFourSelection);
                 }
 
-                origin = ShapesUtil.GetCenterPoint(selectedShapes[1]);
+                origin = ShapeUtil.GetCenterPoint(selectedShapes[1]);
                 startingAngle = (float)AngleBetweenTwoPoints(origin, GetVisualCenter(selectedShapes[2]));
                 var endingAngle = (float)AngleBetweenTwoPoints(origin, GetVisualCenter(selectedShapes[3]));
 
@@ -2067,7 +2067,7 @@ namespace PowerPointLabs.PositionsLab
 
             for (int i = startingIndex; i <= selectedShapes.Count; i++)
             {
-                var angle = (float)AngleBetweenTwoPoints(origin, ShapesUtil.GetCenterPoint(selectedShapes[i]));
+                var angle = (float)AngleBetweenTwoPoints(origin, ShapeUtil.GetCenterPoint(selectedShapes[i]));
                 var angleFromStart = (angle + (360 - startingAngle)) % 360;
 
                 var shapeAngleInfo = new ShapeAngleInfo(selectedShapes[i], angleFromStart);
@@ -2195,7 +2195,7 @@ namespace PowerPointLabs.PositionsLab
                 }
                 else
                 {
-                    sortedShapes = ShapesUtil.SortShapesByLeft(selectedShapes);
+                    sortedShapes = ShapeUtil.SortShapesByLeft(selectedShapes);
                 }
             }
             else
@@ -2216,7 +2216,7 @@ namespace PowerPointLabs.PositionsLab
                     var nextPos = GetSwapReferencePoint(sortedShapes[i + 1], PositionsLabSettings.SwapReferencePoint);
                     currentShape.IncrementLeft(nextPos.X - currentPos.X);
                     currentShape.IncrementTop(nextPos.Y - currentPos.Y);
-                    ShapesUtil.SwapZOrder(currentShape._shape, sortedShapes[i + 1]._shape);
+                    ShapeUtil.SwapZOrder(currentShape._shape, sortedShapes[i + 1]._shape);
                 }
                 else
                 {
@@ -2245,7 +2245,7 @@ namespace PowerPointLabs.PositionsLab
         
         public static void Rotate(Shape shape, Drawing.PointF origin, float angle, PositionsLabSettings.RadialShapeOrientationObject shapeOrientation)
         {
-            var unrotatedCenter = ShapesUtil.GetCenterPoint(shape);
+            var unrotatedCenter = ShapeUtil.GetCenterPoint(shape);
             var rotatedCenter = GraphicsUtil.RotatePoint(unrotatedCenter, origin, angle);
 
             shape.Left += (rotatedCenter.X - unrotatedCenter.X);
@@ -2284,14 +2284,14 @@ namespace PowerPointLabs.PositionsLab
                 throw new Exception(ErrorMessageFewerThanTwoSelection);
             }
 
-            var refShapeCenter = ShapesUtil.GetCenterPoint(shapes[0]);
+            var refShapeCenter = ShapeUtil.GetCenterPoint(shapes[0]);
             var isAllSameDir = true;
             var lastDir = -1;
 
             for (var i = 1; i < shapes.Count; i++)
             {
                 var shape = shapes[i];
-                var shapeCenter = ShapesUtil.GetCenterPoint(shape);
+                var shapeCenter = ShapeUtil.GetCenterPoint(shape);
                 var angle = (float) AngleBetweenTwoPoints(refShapeCenter, shapeCenter);
 
                 var dir = GetDirectionWrtRefShape(shape, angle);
@@ -2320,7 +2320,7 @@ namespace PowerPointLabs.PositionsLab
             for (var i = 1; i < shapes.Count; i++)
             {
                 var shape = shapes[i];
-                var shapeCenter = ShapesUtil.GetCenterPoint(shape);
+                var shapeCenter = ShapeUtil.GetCenterPoint(shape);
                 var angle = (float) AngleBetweenTwoPoints(refShapeCenter, shapeCenter);
 
                 float defaultUpAngle = 0;
