+++ PowerPointLabs/PowerPointLabs/ShortcutsLab/ConvertToPicture.cs
@@ -15,7 +15,7 @@ namespace PowerPointLabs.ShortcutsLab
 #pragma warning disable 0618
         public static void Convert(PowerPoint.Selection selection)
         {
-            if (IsSelectionShapeOrText(selection))
+            if (ShapeUtil.IsSelectionShapeOrText(selection))
             {
                 var shape = GetShapeFromSelection(selection);
                 shape = CutPasteShape(shape);
@@ -29,7 +29,7 @@ namespace PowerPointLabs.ShortcutsLab
 
         public static void ConvertAndSave(PowerPoint.Selection selection, string fileName)
         {
-            if (IsSelectionShapeOrText(selection))
+            if (ShapeUtil.IsSelectionShapeOrText(selection))
             {
                 if (selection.HasChildShapeRange)
                 {
@@ -95,25 +95,6 @@ namespace PowerPointLabs.ShortcutsLab
         {
             PowerPoint.Shape result = shapeRange.Count > 1 ? shapeRange.Group() : shapeRange[1];
             return result;
-        }
-
-        private static bool IsSelectionShapeOrText(PowerPoint.Selection selection)
-        {
-            if (selection.Type != PowerPoint.PpSelectionType.ppSelectionShapes &&
-                selection.Type != PowerPoint.PpSelectionType.ppSelectionText)
-            {
-                return false;
-            }
-
-            foreach (PowerPoint.Shape shape in selection.ShapeRange)
-            {
-                if (shape.Type == Office.MsoShapeType.msoPlaceholder)
-                {
-                    return false;
-                }
-            }
-
-            return true;
         }
     }
 }
