+++ PowerPointLabs/PowerPointLabs/Utils/GraphicsUtil.cs
@@ -1,11 +1,9 @@
 ﻿using System;
-using System.Collections.Generic;
 using System.Diagnostics.CodeAnalysis;
 using System.Drawing;
 using System.Drawing.Drawing2D;
 using System.Drawing.Imaging;
 using System.IO;
-using System.Linq;
 using System.Windows.Forms;
 using System.Windows.Media;
 using System.Windows.Media.Imaging;
@@ -29,7 +27,7 @@ namespace PowerPointLabs.Utils
     {
 #pragma warning disable 0618
 
-        #region Const
+        #region Constants
         private static readonly Object FileLock = new object();
         public const float PictureExportingRatio = 96.0f / 72.0f;
         private const float TargetDpi = 96.0f;
@@ -140,122 +138,6 @@ namespace PowerPointLabs.Utils
         public static void ExportSlide(PowerPointSlide slide, string exportPath, float magnifyRatio = 1.0f)
         {
             ExportSlide(slide.GetNativeSlide(), exportPath, magnifyRatio);
-        }
-
-        /// <summary>
-        /// Sort by increasing index.
-        /// </summary>
-        public static void SortByIndex(List<PowerPointSlide> slides)
-        {
-            slides.Sort((sh1, sh2) => sh1.Index - sh2.Index);
-        }
-
-        /// <summary>
-        /// Used for the SquashSlides method.
-        /// This struct holds transition information for an effect.
-        /// </summary>
-        private struct EffectTransition
-        {
-            private readonly MsoAnimTriggerType slideTransition;
-            private readonly float transitionTime;
-
-            public EffectTransition(MsoAnimTriggerType slideTransition, float transitionTime)
-            {
-                this.slideTransition = slideTransition;
-                this.transitionTime = transitionTime;
-            }
-
-            public void ApplyTransition(Effect effect)
-            {
-                effect.Timing.TriggerType = slideTransition;
-                effect.Timing.TriggerDelayTime = transitionTime;
-            }
-        }
-
-        /// <summary>
-        /// Merges multiple animated slides into a single slide.
-        /// TODO: Test this method more thoroughly, in places other than autozoom.
-        /// </summary>
-        public static void SquashSlides(IEnumerable<PowerPointSlide> slides)
-        {
-            PowerPointSlide firstSlide = null;
-            ShapeRange previousShapes = null;
-            EffectTransition slideTransition = new EffectTransition();
-
-            foreach (var slide in slides)
-            {
-                if (firstSlide == null)
-                {
-                    firstSlide = slide;
-                    slideTransition = GetTransitionFromSlide(slide);
-
-                    firstSlide.Transition.AdvanceOnClick = MsoTriState.msoTrue;
-                    firstSlide.Transition.AdvanceOnTime = MsoTriState.msoFalse;
-                    
-                    previousShapes = ShapesUtil.GetShapesWhenTypeNotMatches(firstSlide, firstSlide.Shapes.Range(), MsoShapeType.msoPlaceholder);
-                    continue;
-                }
-
-                var effectSequence = firstSlide.GetNativeSlide().TimeLine.MainSequence;
-                int effectStartIndex = effectSequence.Count + 1;
-                
-                slide.DeleteIndicator();
-                ShapeRange newShapeRange = firstSlide.CopyShapesToSlide(slide.Shapes.Range());
-                newShapeRange.ZOrder(MsoZOrderCmd.msoSendToBack);
-                
-                foreach (Shape shape in newShapeRange)
-                {
-                    AddAppearAnimation(shape, firstSlide, effectStartIndex);
-                }
-                foreach (Shape shape in previousShapes)
-                {
-                    AddDisappearAnimation(shape, firstSlide, effectStartIndex);
-                }
-                slideTransition.ApplyTransition(effectSequence[effectStartIndex]);
-                
-                previousShapes = newShapeRange;
-                slideTransition = GetTransitionFromSlide(slide);
-                slide.Delete();
-            }
-        }
-
-        /// <summary>
-        /// Extracts the transition animation out of slide to be used as a transition animation for shapes.
-        /// For now, it only extracts the trigger type (trigger by wait or by mouse click), not actual slide transitions.
-        /// </summary>
-        private static EffectTransition GetTransitionFromSlide(PowerPointSlide slide)
-        {
-            var transition = slide.GetNativeSlide().SlideShowTransition;
-            
-            if (transition.AdvanceOnTime == MsoTriState.msoTrue)
-            {
-                return new EffectTransition(MsoAnimTriggerType.msoAnimTriggerAfterPrevious, transition.AdvanceTime);
-            }
-            return new EffectTransition(MsoAnimTriggerType.msoAnimTriggerOnPageClick, 0);
-        }
-
-        private static void AddDisappearAnimation(Shape shape, PowerPointSlide inSlide, int effectStartIndex)
-        {
-            if (inSlide.HasExitAnimation(shape))
-            {
-                return;
-            }
-
-            var effectFade = inSlide.GetNativeSlide().TimeLine.MainSequence.AddEffect(shape, MsoAnimEffect.msoAnimEffectAppear,
-                MsoAnimateByLevel.msoAnimateLevelNone, MsoAnimTriggerType.msoAnimTriggerWithPrevious, effectStartIndex);
-            effectFade.Exit = MsoTriState.msoTrue;
-        }
-
-        private static void AddAppearAnimation(Shape shape, PowerPointSlide inSlide, int effectStartIndex)
-        {
-            if (inSlide.HasEntryAnimation(shape))
-            {
-                return;
-            }
-
-            var effectFade = inSlide.GetNativeSlide().TimeLine.MainSequence.AddEffect(shape, MsoAnimEffect.msoAnimEffectAppear,
-                MsoAnimateByLevel.msoAnimateLevelNone, MsoAnimTriggerType.msoAnimTriggerWithPrevious, effectStartIndex);
-            effectFade.Exit = MsoTriState.msoFalse;
         }
 
         # endregion
@@ -397,7 +279,7 @@ namespace PowerPointLabs.Utils
 
         #endregion
 
-        #region Helper Functions
+        #region Helper Methods
         private static double CalculateScalingRatio(Size oldSize, Size newSize)
         {
             double scalingRatio;
